First run through of allowing a drop-in homeserver configuration from vanilla Synapse docker image.
This part should allow parsing of a given configuration to find listener blocks and use the port numbers in the nginx configuration to forward back to Synapse.
Also update nginx.conf.j2 for clean up and allowing pass in of the port numbers from the original listeners. Clean up homeserver template as well and
add some more comments.