Next stage: Prepare for potentially separated client and federation listeners.
Begin laying foundation for separated Synapse resource listeners. I'm told this will have no benefit, but may be something found in the wild.
As this is a self-contained docker image, we may force this to be a single listener if found separately, but leave the possibility open. Also,
go ahead and upgrade the proxy connection for the main process to HTTP 1.1 with a basic keepalive setting while here since this file got it's
revamp early. It really did need a bit of a clean up.