Next stage: part 3. Replication listener.
Originally, the synapse-workers docker image declared that the replication listener was set to 9093. We need to
be able to deal with drop-in homeserver.yaml files now, so any given homeserver.yaml may have had an existing
'replication' listener already declared that now we don't want to use. Opinion: unilaterally only use 9093 as
the port number and log that the original listener is being ignored. An exposed replication listener(for instance,
on the same port as 'client' or 'federation') would be externally accessable. That could be catastrophic.

Note: this only applies to the HTTP replication listener and has nothing to do with redis replication.

Potentially in the future, might look into being able to accept an external set of workers either from the original
docker image or from this image with no master process but a grouping of workers requested. At that point, will need
to revisit handling of replication listeners.