Allow some worker types to be load balanced by room name (#19)
* Use Tom's awesome regex to derive the room name from a URL path

* Split the client_reader worker into two type, room and non_room

* Split the federation_reader into two types, room and non_room

Also added a few notes and two additional endpoints that were proven to work on workers

* ROLES* move, split shorthand combos, update notes

The shorthand allows for the historical client_reader and federation_reader workers to still work
as options by substituting the combinations they are comprised of when parsing

* The client endpoint for /directory/list/room.* does not work on workers yet

* Adjust processing the shorthand from the requested worker types to be recursive

Then if a requested combined worker is compromised of a shorthand combo worker, it will be handled too

* Break the method to define a load balancing hash out of from it's location

Keeps the processing logic out of the main block

* For worker roles that are not all in the LB block, just use round-robin as the load balance

This should emulate the prior behavior. See comment added

* Create new room_worker role that can be nginx routed based on hash of room name

* Allow duplicate worker subtypes to be merged on a given worker

This allows combinations such as `client_reader+room_worker` to co-exist by merging
the subtypes of `client_reader_room`

* Update file comments at top

* Create new general_reader to pair with room_worker, in that it will handle all other reader functions that are not for room_worker's

* Adjust tests

Remove three tests that are largely redundant now. They work fine and contribute nothing

Last custom test was changed to use the new room_worker and general_reader worker

* New configuration options for controlling behavior around upstream server fails

* Stop forcing the tuning of socket keepalive options, just use the host system defaults

* fix test parameter for shorthand worker type that doesn't exist anymore(renamed it)